requirements_azure_pipeline.txt
# Healthcare Data Pipeline Requirements
# Azure Document Intelligence + PostgreSQL Integration

# Existing requirements (keep these)
selenium==4.15.0
beautifulsoup4==4.12.2
requests==2.31.0
webdriver-manager==4.0.1
pandas==2.1.3
numpy==1.25.2
lxml==4.9.3
python-dateutil==2.8.2
python-dotenv==1.0.0
tqdm==4.66.1
pathlib2==2.3.7
PyPDF2==3.0.1
PyMuPDF==1.23.8
regex==2023.10.3
urllib3==2.1.0
ujson==5.8.0

# Azure Document Intelligence
azure-ai-documentintelligence==1.0.0b1
azure-core==1.29.5
azure-identity==1.15.0

# PostgreSQL Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Async processing
aiofiles==23.2.1
asyncio-throttle==1.0.2

# Data validation and serialization
pydantic==2.5.0
marshmallow==3.20.1

# Environment and configuration
python-dotenv==1.0.0
click==8.1.7

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0